Moved get_rollen to handlers.py

+++ backend/src/zac/core/cache.py
@@ -5,6 +5,7 @@ from django.core.cache import cache, caches
 
 from zgw_consumers.api_models.constants import VertrouwelijkheidsAanduidingen
 from zgw_consumers.api_models.documenten import Document
+from zgw_consumers.api_models.zaken import Rol
 from zgw_consumers.client import Client
 
 from zgw.models.zrc import Zaak
@@ -71,16 +72,14 @@ def invalidate_document_cache(document: Document):
     cache.delete_many(keys)
 
 
-def invalidate_rollen_cache(zaak: Zaak):
+def invalidate_rollen_cache(zaak: Zaak, rollen: Optional[List[Rol]] = None):
     _cache = caches["request"]
     if _cache:
         _cache.delete(f"rollen:{zaak.url}")
 
-    from zac.core.services import get_rollen
+    if rollen:
+        cache_keys = []
+        for rol in rollen:
+            cache_keys.append(f"rol:{rol.url}")
 
-    rollen = get_rollen(zaak)
-    cache_keys = []
-    for rol in rollen:
-        cache_keys.append(f"rol:{rol.url}")
-
-    cache.delete_many(cache_keys)
+        cache.delete_many(cache_keys)
